# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/VivekVinushanth/DataIngestion/1.0.0
info:
  description: API for ingesting profile events and personality data
  version: "1.0.0"
  title: Data Ingestion API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
paths:
...
  /profile:
    post:
      summary: Create a user profile
      operationId: createProfile
      description: Merge different identities of a known user to unify the profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileData'
      responses:
        '201':
          description: Alias successfully updated
        '400':
          description: Invalid alias request
...
  /{perma_id}/profile/personality:
    get:
      summary: Fetch 360 profile of a user
      operationId: getPersonalityProfileData
      description: Retrieve profile of a specific user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
      responses:
        '200':
          description: Personality data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalityData'
        '400':
          description: Bad input parameter

    patch:
      summary: update personality data for a user
      operationId: addPersonalityData
      description: Stores personality data linked to a user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalityData'
      responses:
        '201':
          description: Personality data added successfully
        '400':
          description: Invalid input data

    put:
      summary: update personality data for a user
      operationId: addPersonalityData
      description: Stores personality data linked to a user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalityData'
      responses:
        '201':
          description: Personality data added successfully
        '400':
          description: Invalid input data
  /{perma_id}/profile/app_context:
    get:
      summary: Fetch App context of the user
      operationId: getListofAppContextData
      description: Retrieve profile of a specific user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
      responses:
        '200':
          description: Personality data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppContext'
        '400':
          description: Bad input parameter
  /{perma_id}/profile/{app_id}/app_context:
    get:
      summary: Fetch App context of the user
      operationId: getAppContextData
      description: Retrieve profile of a specific user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
      responses:
        '200':
          description: Personality data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppContext'
        '400':
          description: Bad input parameter
    patch:
      summary: update personality data for a user
      operationId: addAppContextData
      description: Stores personality data linked to a user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppContext'
      responses:
        '201':
          description: Personality data added successfully
        '400':
          description: Invalid input data

    put:
      summary: update personality data for a user
      operationId: addAppContextData
      description: Stores personality data linked to a user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppContext'
      responses:
        '201':
          description: Personality data added successfully
        '400':
          description: Invalid input data
...
components:
  schemas:
...
    ProfileData:
      type: object
      required:
        - originCountry
      properties:
        originCountry:
          type: string
          example: LK
        user_ids:
          type: array
          items:
            type: string
            format: uuid
        identity:
          $ref: '#/components/schemas/IdentityData'
        personality:
          $ref: '#/components/schemas/PersonalityData'
        app_context:
          type: array
          items:
            $ref: '#/components/schemas/AppContext'
...
    IdentityData:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone_numbers:
          type: array
          items:
            type: string
        first_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        preferred_username:
          type: string
        profile_url:
          type: string
          format: uri
        picture:
          type: string
          format: uri
        roles:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        account_status:
          type: string
          enum: [active, inactive, suspended]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        idp_provider:
          type: string
        mfa_enabled:
          type: boolean
        last_login:
          type: string
          format: date-time
        locale:
          type: string
        timezone:
          type: string
    PersonalityData:
      type: object
      properties:
        interests:
          type: array
          items:
            type: string
        preferred_language:
          type: string
        communication_preferences:
          type: object
          properties:
            email_notifications:
              type: boolean
            sms_notifications:
              type: boolean
            push_notifications:
              type: boolean
        shopping_preferences:
          type: object
          properties:
            favorite_brands:
              type: array
              items:
                type: string
            discount_preference:
              type: string
    AppContext:
      type: object
      properties:
        app_id:
          type: string
        subscription_plan:
          type: string
        app_permissions:
          type: array
          items:
            type: string
        feature_flags:
          type: object
          properties:
            beta_features_enabled:
              type: boolean
            dark_mode:
              type: boolean
        last_active_app:
          type: string
        usage_metrics:
          type: object
          properties:
            daily_active_time:
              type: integer
            monthly_logins:
              type: integer
        devices:
          type: array
          items:
            type: object
            properties:
              device_id:
                type: string
              device_type:
                type: string
              os:
                type: string
              browser:
                type: string
              browser_version:
                type: string
              ip:
                type: string
              last_used:
                type: string
                format: date-time
        regions_accessed:
          type: array
          items:
            type: object
            properties:
              country:
                type: string
              city:
                type: string
              timezone:
                type: string
              first_accessed:
                type: string
                format: date-time
              last_accessed:
                type: string
                format: date-time
...